pyproject.toml
[project]
name = "pdf-loader-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.2",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "pdfminer-six>=20260107",
    "pi-heif>=1.3.0",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.55.0",
    "unstructured>=0.21.5",
    "unstructured-inference>=1.2.0",
]
